chore: Sample upgrade
